Update export so we can use it on CP without problem and warnings
