+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "@ananasgit/ananas-fe-shared",
-  "version": "1.1.0",
+  "version": "1.1.2",
   "main": "dist/index.js",
+  "types": "dist/index.ts",
   "files": [
     "dist"
   ],
-  "type": "module",
   "author": "ananas",
   "repository": {
     "type": "git",
@@ -20,7 +20,6 @@
     "prettier": "^3.3.3",
     "typescript": "5.5.4"
   },
-  "types": "src/index.ts",
   "scripts": {
     "build": "tsc",
     "lint": "eslint .",
